+++ rest_framework_services/types/__init__.py
@@ -7,6 +7,7 @@ inputs/outputs are not coupled to the mutation helpers themselves.
 
 from rest_framework_services.types.change_result import ChangeResult
 from rest_framework_services.types.field_change import FieldChange
+from rest_framework_services.types.http_extras import HttpExtras
 from rest_framework_services.types.no_input import NoInput
 from rest_framework_services.types.no_kwargs import NoKwargs
 from rest_framework_services.types.selector_spec import SelectorSpec
@@ -18,6 +19,7 @@ __all__ = [
     "UNSET",
     "ChangeResult",
     "FieldChange",
+    "HttpExtras",
     "NoInput",
     "NoKwargs",
     "SelectorSpec",